Moved endToEndTest() to MainAcitivituTest class. Removed step to explicitely set the initial direction to North.
