+++ app/src/androidTest/java/com/knthcame/marsrover/MainActivityTest.kt
@@ -1,10 +1,16 @@
 package com.knthcame.marsrover
 
+import androidx.compose.ui.test.ExperimentalTestApi
 import androidx.compose.ui.test.assertIsDisplayed
+import androidx.compose.ui.test.assertTextEquals
+import androidx.compose.ui.test.hasTestTag
 import androidx.compose.ui.test.junit4.createAndroidComposeRule
 import androidx.compose.ui.test.onNodeWithTag
+import androidx.compose.ui.test.performClick
+import androidx.compose.ui.test.performTextReplacement
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
+import com.knthcame.marsrover.ui.movements.Movement
 import org.junit.Assert.assertEquals
 import org.junit.Rule
 import org.junit.Test
@@ -31,5 +37,62 @@ class MainActivityTest {
     fun testLaunch() {
         composeActivityRule.onNodeWithTag("homeTopBarTitle", useUnmergedTree = true)
             .assertIsDisplayed()
+    }
+
+    @OptIn(ExperimentalTestApi::class)
+    @Test
+    fun endToEndTest() {
+        // Navigate to setup screen
+        composeActivityRule.onNodeWithTag("homeStartButton")
+            .performClick()
+
+        // Input setup data
+        composeActivityRule.onNodeWithTag("setupPlateauWidthTextField")
+            .performTextReplacement("5")
+        composeActivityRule.onNodeWithTag("setupPlateauHeightTextField")
+            .performTextReplacement("3")
+        composeActivityRule.onNodeWithTag("setupInitialXTextField")
+            .performTextReplacement("1")
+        composeActivityRule.onNodeWithTag("setupInitialYTextField")
+            .performTextReplacement("2")
+        composeActivityRule.onNodeWithTag("setupInitialDirectionTextField")
+            .performClick()
+
+        // Navigate to movements screen
+        composeActivityRule.onNodeWithTag("setupContinueButton")
+            .performClick()
+
+        // Input movements
+        composeActivityRule.onNodeWithTag("add${Movement.RotateLeft}MovementButton")
+            .performClick()
+        composeActivityRule.onNodeWithTag("add${Movement.MoveForward}MovementButton")
+            .performClick()
+        composeActivityRule.onNodeWithTag("add${Movement.RotateLeft}MovementButton")
+            .performClick()
+        composeActivityRule.onNodeWithTag("add${Movement.MoveForward}MovementButton")
+            .performClick()
+        composeActivityRule.onNodeWithTag("add${Movement.RotateLeft}MovementButton")
+            .performClick()
+        composeActivityRule.onNodeWithTag("add${Movement.MoveForward}MovementButton")
+            .performClick()
+        composeActivityRule.onNodeWithTag("add${Movement.RotateLeft}MovementButton")
+            .performClick()
+        composeActivityRule.onNodeWithTag("add${Movement.MoveForward}MovementButton")
+            .performClick()
+        composeActivityRule.onNodeWithTag("add${Movement.MoveForward}MovementButton")
+            .performClick()
+
+        // Send instructions
+        composeActivityRule.onNodeWithTag("sendMovementsButton")
+            .performClick()
+
+        // Assert correct output in alert dialog.
+        val outputTextTestTag = "movementsOutputDialogOutputText"
+        composeActivityRule.waitUntilExactlyOneExists(
+            hasTestTag(outputTextTestTag),
+            timeoutMillis = 5_000
+        )
+        composeActivityRule.onNodeWithTag(outputTextTestTag)
+            .assertTextEquals("1 3 N")
     }
 }
